requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
streamlit==1.55.0
streamlit-folium==0.27.1
folium==0.20.0
requests==2.32.5
pandas==2.3.2
numpy==2.4.4
scipy==1.17.1
scikit-learn==1.8.0
python-dotenv==1.1.1
